Changed to only use ints from uints

+++ src/push_swap.c
@@ -18,7 +18,7 @@ int	main(int argc, char *argv[])
 
 	if (input_check(argc, argv) != 0) // TODO sort out input check
 		exit(0);
-	var.n_a = (unsigned)argc - 1; // TODO input as a single string
+	var.n_a = argc - 1; // TODO input as a single string
 	var.args = parse_args(argc, argv); // TODO input as a single string
 	hash_array(&var);
 	list_moves(&var);
